+++ packages/did-ui-react/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.0-alpha.19](https://github.com/Portkey-Wallet/portkey-web/compare/v2.0.0-alpha.18...v2.0.0-alpha.19) (2024-01-30)
+
+**Note:** Version bump only for package @portkey/did-ui-react
+
+## [2.0.0-alpha.18](https://github.com/Portkey-Wallet/portkey-web/compare/v2.0.0-alpha.17...v2.0.0-alpha.18) (2024-01-30)
+
+### Bug Fixes
+
+- 🐛 qr data ([f4f333b](https://github.com/Portkey-Wallet/portkey-web/commit/f4f333b5666da60a07313b98f02b9ae81e1f2ada))
+
+## [2.0.0-alpha.17](https://github.com/Portkey-Wallet/portkey-web/compare/v2.0.0-alpha.16...v2.0.0-alpha.17) (2024-01-29)
+
+### Features
+
+- 🎸 add finish resutl create type ([b5ab726](https://github.com/Portkey-Wallet/portkey-web/commit/b5ab726364d70ee363a9c80d8560574570cdeb7d))
+
 ## [2.0.0-alpha.16](https://github.com/Portkey-Wallet/portkey-web/compare/v2.0.0-alpha.15...v2.0.0-alpha.16) (2024-01-26)
 
 ### Bug Fixes